Don't allow duplicate slugs
Using the pages available in the current page list, do not allow duplicate slugs on the page editor. Disable the save action when duplicate slugs are detected.

Note: we might need to make this an async validation if we get to the point where the page list isn't showing all of the pages that should be validated against. (whenever we add search/filtering back in)