Remove media plugin
For now, I am going to remove this plugin. There isn’t a good way to modify it to work the way we want. If anything we can try and create our own custom embed plugin - but we might not have time for that in MVP. 

In the meantime, users who can edit source control will be able to embed videos by pasting in the embed code from youtube and other embeddable content sites.